ovirtlago: drop ovirt.collect tests
This patch drops the functional tests for 'lago ovirt collect',
as this command is deprecated since v.0.30, and just redirects to
'lago collect'.  We already test 'collect' functionality
in 'collect.bats'.

One testing functionality that is dropped here, is checking
the 'inherited_artifacts' in oVirt VM types.
I don't think this test justifies starting a new (fc23)
VM for that. Also it looses its relevance as we collect the
entire '/var/log' directory now.

Signed-off-by: Nadav Goldin <ngoldin@redhat.com>