lago ovirt start: Match cluster CPU compatibility
When exporting images and starting them on a different host, Lago will
set the VM's CPU type to libvirt 'host-passthrough' mode by default. This
means that when the VMs start, the CPU type of the host(in oVirt's eyes)
could have changed without it being aware. So this patch ensures the
cluster compatibility level would be matched to the new CPU, if it
changed,  when running 'lago ovirt start'.

Signed-off-by: Nadav Goldin <ngoldin@redhat.com>