Download Engine CA certificte
For the time being, we can't use it: we connect to Engine using IP
and the certificate is signed for 'engine'. When we switch from IP
to DNS, it could work, but this should be done in a separate commit.

This will be useful when we wish to use the certificate in some tests
(image upload, OVN API tests, etc.)